app/requirements.txt
# Google ADK
google-adk>=1.21.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0

# HTTP Client
requests>=2.31.0
aiohttp>=3.9.0

# Numerical Calculations
numpy>=1.24.0

# CSV Processing
pandas>=2.0.0
